Move wlayout to wcore, create new tab layout for all new tabs
